requirements.txt
nornir
nornir-jinja2
nornir-scrapli
nornir-utils
pybatfish
rich
python-dotenv